requirements.txt
# Format code
autoflake==2.3.1
black==24.10.0
isort==5.13.2

# Dependencies
Flask==3.1.0
Flask-Alembic==3.1.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
geopandas==1.0.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
